pyproject.toml
[build-system]
requires      = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name         = "dbqq"
version      = "0.4.6"
dependencies = [
    "rsa",
    "pyaml",
    "polars",
    "databricks-sql-connector",
    "pyarrow",
    "connectorx",
    "triple-quote-clean"
]
authors = [
    {name = "Chris Mamon", email="chrisam1993@live.com"}
]

[project.scripts]
dbqq-write-keys = "dbqq.security._cli_write_keys:cli_write_keys"
dbqq-encrypt-yaml = "dbqq.security._cli_encrypt_yaml:cli_encrypt_yaml"